+++ pom.xml
@@ -55,7 +55,7 @@
   <org.apache.commons.io.version>2.6</org.apache.commons.io.version>
   <org.apache.httpcomponents.version>4.5.11</org.apache.httpcomponents.version>
   <io.fabric8.openshift.client.version>4.6.4</io.fabric8.openshift.client.version>
-  <io.kubernetes.client.java.version>5.0.0</io.kubernetes.client.java.version>
+  <io.fabric8.kubernetes.client.version>4.6.4</io.fabric8.kubernetes.client.version>
   <io.vertx.web.version>3.9.2</io.vertx.web.version>
   <org.slf4j.version>1.7.30</org.slf4j.version>
   <com.google.code.gson.version>2.8.6</com.google.code.gson.version>
@@ -84,9 +84,9 @@
     <version>${io.fabric8.openshift.client.version}</version>
   </dependency>
   <dependency>
-    <groupId>io.kubernetes</groupId>
-    <artifactId>client-java</artifactId>
-    <version>${io.kubernetes.client.java.version}</version>
+    <groupId>io.fabric8</groupId>
+    <artifactId>kubernetes-client</artifactId>
+    <version>${io.fabric8.kubernetes.client.version}</version>
   </dependency>
   <dependency>
     <groupId>com.google.dagger</groupId>
