+++ src/main/java/com/redhat/rhjmc/containerjfr/platform/ServiceRef.java
@@ -46,8 +46,8 @@ import java.util.Optional;
 
 import javax.management.remote.JMXServiceURL;
 
-import org.apache.commons.lang.builder.EqualsBuilder;
-import org.apache.commons.lang.builder.HashCodeBuilder;
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import com.google.gson.annotations.SerializedName;
