+++ src/test/java/com/redhat/rhjmc/containerjfr/platform/internal/KubeApiPlatformClientTest.java
@@ -41,23 +41,20 @@
  */
 package com.redhat.rhjmc.containerjfr.platform.internal;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static org.mockito.Mockito.when;
-
-import java.net.MalformedURLException;
-import java.net.UnknownHostException;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 import javax.management.remote.JMXServiceURL;
 
+import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.InOrder;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.invocation.InvocationOnMock;
@@ -66,104 +63,123 @@ import org.mockito.stubbing.Answer;
 
 import com.redhat.rhjmc.containerjfr.core.log.Logger;
 import com.redhat.rhjmc.containerjfr.core.net.JFRConnectionToolkit;
+import com.redhat.rhjmc.containerjfr.core.net.discovery.JvmDiscoveryClient.EventKind;
+import com.redhat.rhjmc.containerjfr.core.sys.Environment;
+import com.redhat.rhjmc.containerjfr.messaging.notifications.Notification;
+import com.redhat.rhjmc.containerjfr.messaging.notifications.NotificationFactory;
 import com.redhat.rhjmc.containerjfr.platform.ServiceRef;
 
-import io.kubernetes.client.ApiException;
-import io.kubernetes.client.apis.CoreV1Api;
-import io.kubernetes.client.models.V1EndpointAddress;
-import io.kubernetes.client.models.V1EndpointPort;
-import io.kubernetes.client.models.V1EndpointSubset;
-import io.kubernetes.client.models.V1Endpoints;
-import io.kubernetes.client.models.V1EndpointsList;
-import io.kubernetes.client.models.V1ObjectReference;
+import io.fabric8.kubernetes.api.model.EndpointAddress;
+import io.fabric8.kubernetes.api.model.EndpointPort;
+import io.fabric8.kubernetes.api.model.EndpointSubset;
+import io.fabric8.kubernetes.api.model.Endpoints;
+import io.fabric8.kubernetes.api.model.EndpointsList;
+import io.fabric8.kubernetes.api.model.ObjectReference;
+import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.Watcher;
+import io.fabric8.kubernetes.client.Watcher.Action;
+import io.fabric8.kubernetes.client.dsl.MixedOperation;
+import io.fabric8.kubernetes.client.dsl.NonNamespaceOperation;
 
 @ExtendWith(MockitoExtension.class)
 class KubeApiPlatformClientTest {
 
-    KubeApiPlatformClient client;
-    @Mock CoreV1Api api;
-    String namespace = "someNamespace";
+    static final String NAMESPACE = "foo-namespace";
+
+    KubeApiPlatformClient platformClient;
+    @Mock KubernetesClient k8sClient;
     @Mock JFRConnectionToolkit connectionToolkit;
+    @Mock Environment env;
+    @Mock NotificationFactory notificationFactory;
     @Mock Logger logger;
 
     @BeforeEach
-    void setup() {
-        client = new KubeApiPlatformClient(api, namespace, () -> connectionToolkit, logger);
+    void setup() throws Exception {
+        this.platformClient =
+                new KubeApiPlatformClient(
+                        NAMESPACE, k8sClient, () -> connectionToolkit, notificationFactory, logger);
     }
 
     @Test
-    void discoversNoServicesIfApiThrows() throws ApiException {
-        when(api.listNamespacedEndpoints(
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any()))
-                .thenThrow(ApiException.class);
+    void shouldReturnEmptyListIfNoEndpointsFound() throws Exception {
+        MixedOperation mockNamespaceOperation = Mockito.mock(MixedOperation.class);
+        Mockito.when(k8sClient.endpoints()).thenReturn(mockNamespaceOperation);
 
-        assertThat(client.listDiscoverableServices(), Matchers.empty());
+        NonNamespaceOperation mockOperation = Mockito.mock(NonNamespaceOperation.class);
+        Mockito.when(mockNamespaceOperation.inNamespace(Mockito.anyString()))
+                .thenReturn(mockOperation);
 
-        verify(api)
-                .listNamespacedEndpoints(
-                        namespace, null, null, null, null, null, null, null, null, null);
-        verifyNoMoreInteractions(api);
+        EndpointsList mockListable = Mockito.mock(EndpointsList.class);
+        Mockito.when(mockOperation.list()).thenReturn(mockListable);
+
+        List<Endpoints> mockEndpoints = Collections.emptyList();
+        Mockito.when(mockListable.getItems()).thenReturn(mockEndpoints);
+
+        platformClient.start();
+        List<ServiceRef> result = platformClient.listDiscoverableServices();
+        MatcherAssert.assertThat(result, Matchers.equalTo(Collections.emptyList()));
     }
 
     @Test
-    void discoversServices() throws ApiException, UnknownHostException, MalformedURLException {
-        V1EndpointsList mockServiceList = mock(V1EndpointsList.class);
+    void shouldReturnListOfMatchingEndpointRefs() throws Exception {
+        MixedOperation mockNamespaceOperation = Mockito.mock(MixedOperation.class);
+        Mockito.when(k8sClient.endpoints()).thenReturn(mockNamespaceOperation);
 
-        V1Endpoints mockServiceA = mock(V1Endpoints.class);
-        V1EndpointAddress aAddr = mock(V1EndpointAddress.class);
-        when(aAddr.getIp()).thenReturn("127.0.0.1");
-        V1ObjectReference aRef = mock(V1ObjectReference.class);
-        when(aRef.getName()).thenReturn("ServiceA.local");
-        when(aAddr.getTargetRef()).thenReturn(aRef);
-        V1EndpointSubset aSubset = mock(V1EndpointSubset.class);
-        when(aSubset.getAddresses()).thenReturn(List.of(aAddr));
-        when(mockServiceA.getSubsets()).thenReturn(List.of(aSubset));
-        V1EndpointPort aPort1 = mock(V1EndpointPort.class);
-        when(aPort1.getName()).thenReturn("jfr-jmx");
-        when(aPort1.getPort()).thenReturn(123);
-        V1EndpointPort aPort2 = mock(V1EndpointPort.class);
-        when(aPort2.getName()).thenReturn("port-name");
-        when(aPort2.getPort()).thenReturn(456);
-        when(aSubset.getPorts()).thenReturn(List.of(aPort1, aPort2));
+        NonNamespaceOperation mockOperation = Mockito.mock(NonNamespaceOperation.class);
+        Mockito.when(mockNamespaceOperation.inNamespace(Mockito.anyString()))
+                .thenReturn(mockOperation);
 
-        V1Endpoints mockServiceB = mock(V1Endpoints.class);
-        V1EndpointAddress bAddr = mock(V1EndpointAddress.class);
-        when(bAddr.getIp()).thenReturn("10.0.0.1");
-        V1ObjectReference bRef = mock(V1ObjectReference.class);
-        when(bRef.getName()).thenReturn("b-service.example.com");
-        when(bAddr.getTargetRef()).thenReturn(bRef);
-        V1EndpointSubset bSubset = mock(V1EndpointSubset.class);
-        when(bSubset.getAddresses()).thenReturn(List.of(bAddr));
-        when(mockServiceB.getSubsets()).thenReturn(List.of(bSubset));
-        V1EndpointPort bPort1 = mock(V1EndpointPort.class);
-        when(bPort1.getName()).thenReturn("jfr-jmx");
-        when(bPort1.getPort()).thenReturn(7899);
-        when(bSubset.getPorts()).thenReturn(List.of(bPort1));
+        EndpointsList mockListable = Mockito.mock(EndpointsList.class);
+        Mockito.when(mockOperation.list()).thenReturn(mockListable);
 
-        when(mockServiceList.getItems()).thenReturn(Arrays.asList(mockServiceA, mockServiceB));
-        when(api.listNamespacedEndpoints(
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any()))
-                .thenReturn(mockServiceList);
+        ObjectReference objRef1 = Mockito.mock(ObjectReference.class);
+        // Mockito.when(objRef1.getName()).thenReturn("targetA");
+        ObjectReference objRef2 = Mockito.mock(ObjectReference.class);
+        Mockito.when(objRef2.getName()).thenReturn("targetB");
+        ObjectReference objRef3 = Mockito.mock(ObjectReference.class);
+        Mockito.when(objRef3.getName()).thenReturn("targetC");
+        ObjectReference objRef4 = Mockito.mock(ObjectReference.class);
+        Mockito.when(objRef4.getName()).thenReturn("targetD");
 
-        when(connectionToolkit.createServiceURL(Mockito.anyString(), Mockito.anyInt()))
+        EndpointAddress address1 = Mockito.mock(EndpointAddress.class);
+        // Mockito.when(address1.getIp()).thenReturn("127.0.0.1");
+        // Mockito.when(address1.getTargetRef()).thenReturn(objRef1);
+        EndpointAddress address2 = Mockito.mock(EndpointAddress.class);
+        Mockito.when(address2.getIp()).thenReturn("127.0.0.2");
+        Mockito.when(address2.getTargetRef()).thenReturn(objRef2);
+        EndpointAddress address3 = Mockito.mock(EndpointAddress.class);
+        Mockito.when(address3.getIp()).thenReturn("127.0.0.3");
+        Mockito.when(address3.getTargetRef()).thenReturn(objRef3);
+        EndpointAddress address4 = Mockito.mock(EndpointAddress.class);
+        Mockito.when(address4.getIp()).thenReturn("127.0.0.4");
+        Mockito.when(address4.getTargetRef()).thenReturn(objRef4);
+
+        EndpointPort port1 = Mockito.mock(EndpointPort.class);
+        Mockito.when(port1.getPort()).thenReturn(80);
+        Mockito.when(port1.getName()).thenReturn("tcp-80");
+        EndpointPort port2 = Mockito.mock(EndpointPort.class);
+        Mockito.when(port2.getPort()).thenReturn(9999);
+        Mockito.when(port2.getName()).thenReturn("jfr-jmx");
+        EndpointPort port3 = Mockito.mock(EndpointPort.class);
+        Mockito.when(port3.getPort()).thenReturn(9091);
+        Mockito.when(port3.getName()).thenReturn("tcp-9091");
+
+        EndpointSubset subset1 = Mockito.mock(EndpointSubset.class);
+        // Mockito.when(subset1.getAddresses()).thenReturn(Collections.singletonList(address1));
+        Mockito.when(subset1.getPorts()).thenReturn(Collections.singletonList(port1));
+        EndpointSubset subset2 = Mockito.mock(EndpointSubset.class);
+        Mockito.when(subset2.getAddresses()).thenReturn(Arrays.asList(address2, address3));
+        Mockito.when(subset2.getPorts()).thenReturn(Collections.singletonList(port2));
+        EndpointSubset subset3 = Mockito.mock(EndpointSubset.class);
+        Mockito.when(subset3.getAddresses()).thenReturn(Collections.singletonList(address4));
+        Mockito.when(subset3.getPorts()).thenReturn(Collections.singletonList(port3));
+
+        Endpoints endpoint = Mockito.mock(Endpoints.class);
+        Mockito.when(endpoint.getSubsets()).thenReturn(Arrays.asList(subset1, subset2, subset3));
+
+        Mockito.when(mockListable.getItems()).thenReturn(Collections.singletonList(endpoint));
+
+        Mockito.when(connectionToolkit.createServiceURL(Mockito.anyString(), Mockito.anyInt()))
                 .thenAnswer(
                         new Answer<>() {
                             @Override
@@ -178,17 +194,254 @@ class KubeApiPlatformClientTest {
                             }
                         });
 
-        List<ServiceRef> result = client.listDiscoverableServices();
-
-        ServiceRef serv1 = new ServiceRef(connectionToolkit, "127.0.0.1", 123, "ServiceA.local");
+        platformClient.start();
+        List<ServiceRef> result = platformClient.listDiscoverableServices();
+        ServiceRef serv1 =
+                new ServiceRef(
+                        connectionToolkit,
+                        address2.getIp(),
+                        port2.getPort(),
+                        address2.getTargetRef().getName());
         ServiceRef serv2 =
-                new ServiceRef(connectionToolkit, "10.0.0.1", 7899, "b-service.example.com");
+                new ServiceRef(
+                        connectionToolkit,
+                        address3.getIp(),
+                        port2.getPort(),
+                        address3.getTargetRef().getName());
+        ServiceRef serv3 =
+                new ServiceRef(
+                        connectionToolkit,
+                        address4.getIp(),
+                        port3.getPort(),
+                        address4.getTargetRef().getName());
 
-        assertThat(result, Matchers.equalTo(List.of(serv1, serv2)));
+        MatcherAssert.assertThat(result, Matchers.equalTo(Arrays.asList(serv1, serv2, serv3)));
+    }
 
-        verify(api)
-                .listNamespacedEndpoints(
-                        namespace, null, null, null, null, null, null, null, null, null);
-        verifyNoMoreInteractions(api);
+    @Test
+    public void shouldSubscribeWatchWhenStarted() throws Exception {
+        MixedOperation op = Mockito.mock(MixedOperation.class);
+        Mockito.when(k8sClient.endpoints()).thenReturn(op);
+        Mockito.when(op.inNamespace(Mockito.anyString())).thenReturn(op);
+
+        Mockito.verifyNoInteractions(k8sClient);
+
+        platformClient.start();
+
+        InOrder inOrder = Mockito.inOrder(k8sClient, op);
+        inOrder.verify(k8sClient).endpoints();
+        inOrder.verify(op).inNamespace(NAMESPACE);
+        inOrder.verify(op).watch(Mockito.any(Watcher.class));
+        Mockito.verifyNoMoreInteractions(k8sClient);
+        Mockito.verifyNoMoreInteractions(op);
+    }
+
+    @Test
+    public void shouldNotifyOnAsyncAdded() throws Exception {
+        MixedOperation op = Mockito.mock(MixedOperation.class);
+        Mockito.when(k8sClient.endpoints()).thenReturn(op);
+        Mockito.when(op.inNamespace(Mockito.anyString())).thenReturn(op);
+
+        Mockito.when(connectionToolkit.createServiceURL(Mockito.anyString(), Mockito.anyInt()))
+                .thenAnswer(
+                        new Answer<>() {
+                            @Override
+                            public JMXServiceURL answer(InvocationOnMock args) throws Throwable {
+                                String host = args.getArgument(0);
+                                int port = args.getArgument(1);
+                                return new JMXServiceURL(
+                                        "rmi",
+                                        "",
+                                        0,
+                                        "/jndi/rmi://" + host + ":" + port + "/jmxrmi");
+                            }
+                        });
+
+        ObjectReference objRef = Mockito.mock(ObjectReference.class);
+        Mockito.when(objRef.getName()).thenReturn("targetA");
+        EndpointAddress address = Mockito.mock(EndpointAddress.class);
+        Mockito.when(address.getIp()).thenReturn("127.0.0.1");
+        Mockito.when(address.getTargetRef()).thenReturn(objRef);
+        EndpointPort port = Mockito.mock(EndpointPort.class);
+        Mockito.when(port.getPort()).thenReturn(9999);
+        Mockito.when(port.getName()).thenReturn("jfr-jmx");
+        EndpointSubset subset = Mockito.mock(EndpointSubset.class);
+        Mockito.when(subset.getAddresses()).thenReturn(Arrays.asList(address));
+        Mockito.when(subset.getPorts()).thenReturn(Collections.singletonList(port));
+
+        Endpoints endpoints = Mockito.mock(Endpoints.class);
+        Mockito.when(endpoints.getSubsets()).thenReturn(Arrays.asList(subset));
+
+        Notification notification = Mockito.mock(Notification.class);
+        Notification.Builder builder = Mockito.mock(Notification.Builder.class);
+        Mockito.when(builder.metaCategory(Mockito.any())).thenReturn(builder);
+        Mockito.when(builder.message(Mockito.any())).thenReturn(builder);
+        Mockito.when(builder.build()).thenReturn(notification);
+        Mockito.when(notificationFactory.createBuilder()).thenReturn(builder);
+
+        platformClient.start();
+
+        ArgumentCaptor<Watcher> watcherCaptor = ArgumentCaptor.forClass(Watcher.class);
+        Mockito.verify(op).watch(watcherCaptor.capture());
+        Watcher watcher = watcherCaptor.getValue();
+        MatcherAssert.assertThat(watcher, Matchers.notNullValue());
+
+        watcher.eventReceived(Action.ADDED, endpoints);
+
+        ServiceRef serviceRef =
+                new ServiceRef(
+                        connectionToolkit,
+                        address.getIp(),
+                        port.getPort(),
+                        address.getTargetRef().getName());
+
+        Mockito.verify(notificationFactory, Mockito.times(1)).createBuilder();
+        Mockito.verify(builder).metaCategory("TargetJvmDiscovery");
+        Mockito.verify(builder)
+                .message(
+                        Map.of("event", Map.of("kind", EventKind.FOUND, "serviceRef", serviceRef)));
+        Mockito.verify(builder).build();
+        Mockito.verify(notification, Mockito.times(1)).send();
+    }
+
+    @Test
+    public void shouldNotifyOnAsyncDeleted() throws Exception {
+        MixedOperation op = Mockito.mock(MixedOperation.class);
+        Mockito.when(k8sClient.endpoints()).thenReturn(op);
+        Mockito.when(op.inNamespace(Mockito.anyString())).thenReturn(op);
+
+        Mockito.when(connectionToolkit.createServiceURL(Mockito.anyString(), Mockito.anyInt()))
+                .thenAnswer(
+                        new Answer<>() {
+                            @Override
+                            public JMXServiceURL answer(InvocationOnMock args) throws Throwable {
+                                String host = args.getArgument(0);
+                                int port = args.getArgument(1);
+                                return new JMXServiceURL(
+                                        "rmi",
+                                        "",
+                                        0,
+                                        "/jndi/rmi://" + host + ":" + port + "/jmxrmi");
+                            }
+                        });
+
+        ObjectReference objRef = Mockito.mock(ObjectReference.class);
+        Mockito.when(objRef.getName()).thenReturn("targetA");
+        EndpointAddress address = Mockito.mock(EndpointAddress.class);
+        Mockito.when(address.getIp()).thenReturn("127.0.0.1");
+        Mockito.when(address.getTargetRef()).thenReturn(objRef);
+        EndpointPort port = Mockito.mock(EndpointPort.class);
+        Mockito.when(port.getPort()).thenReturn(9999);
+        Mockito.when(port.getName()).thenReturn("jfr-jmx");
+        EndpointSubset subset = Mockito.mock(EndpointSubset.class);
+        Mockito.when(subset.getAddresses()).thenReturn(Arrays.asList(address));
+        Mockito.when(subset.getPorts()).thenReturn(Collections.singletonList(port));
+
+        Endpoints endpoints = Mockito.mock(Endpoints.class);
+        Mockito.when(endpoints.getSubsets()).thenReturn(Arrays.asList(subset));
+
+        Notification notification = Mockito.mock(Notification.class);
+        Notification.Builder builder = Mockito.mock(Notification.Builder.class);
+        Mockito.when(builder.metaCategory(Mockito.any())).thenReturn(builder);
+        Mockito.when(builder.message(Mockito.any())).thenReturn(builder);
+        Mockito.when(builder.build()).thenReturn(notification);
+        Mockito.when(notificationFactory.createBuilder()).thenReturn(builder);
+
+        platformClient.start();
+
+        ArgumentCaptor<Watcher> watcherCaptor = ArgumentCaptor.forClass(Watcher.class);
+        Mockito.verify(op).watch(watcherCaptor.capture());
+        Watcher watcher = watcherCaptor.getValue();
+        MatcherAssert.assertThat(watcher, Matchers.notNullValue());
+
+        watcher.eventReceived(Action.DELETED, endpoints);
+
+        ServiceRef serviceRef =
+                new ServiceRef(
+                        connectionToolkit,
+                        address.getIp(),
+                        port.getPort(),
+                        address.getTargetRef().getName());
+
+        Mockito.verify(notificationFactory, Mockito.times(1)).createBuilder();
+        Mockito.verify(builder).metaCategory("TargetJvmDiscovery");
+        Mockito.verify(builder)
+                .message(Map.of("event", Map.of("kind", EventKind.LOST, "serviceRef", serviceRef)));
+        Mockito.verify(builder).build();
+        Mockito.verify(notification, Mockito.times(1)).send();
+    }
+
+    @Test
+    public void shouldNotifyOnAsyncModified() throws Exception {
+        MixedOperation op = Mockito.mock(MixedOperation.class);
+        Mockito.when(k8sClient.endpoints()).thenReturn(op);
+        Mockito.when(op.inNamespace(Mockito.anyString())).thenReturn(op);
+
+        Mockito.when(connectionToolkit.createServiceURL(Mockito.anyString(), Mockito.anyInt()))
+                .thenAnswer(
+                        new Answer<>() {
+                            @Override
+                            public JMXServiceURL answer(InvocationOnMock args) throws Throwable {
+                                String host = args.getArgument(0);
+                                int port = args.getArgument(1);
+                                return new JMXServiceURL(
+                                        "rmi",
+                                        "",
+                                        0,
+                                        "/jndi/rmi://" + host + ":" + port + "/jmxrmi");
+                            }
+                        });
+
+        ObjectReference objRef = Mockito.mock(ObjectReference.class);
+        Mockito.when(objRef.getName()).thenReturn("targetA");
+        EndpointAddress address = Mockito.mock(EndpointAddress.class);
+        Mockito.when(address.getIp()).thenReturn("127.0.0.1");
+        Mockito.when(address.getTargetRef()).thenReturn(objRef);
+        EndpointPort port = Mockito.mock(EndpointPort.class);
+        Mockito.when(port.getPort()).thenReturn(9999);
+        Mockito.when(port.getName()).thenReturn("jfr-jmx");
+        EndpointSubset subset = Mockito.mock(EndpointSubset.class);
+        Mockito.when(subset.getAddresses()).thenReturn(Arrays.asList(address));
+        Mockito.when(subset.getPorts()).thenReturn(Collections.singletonList(port));
+
+        Endpoints endpoints = Mockito.mock(Endpoints.class);
+        Mockito.when(endpoints.getSubsets()).thenReturn(Arrays.asList(subset));
+
+        Notification notification = Mockito.mock(Notification.class);
+        Notification.Builder builder = Mockito.mock(Notification.Builder.class);
+        Mockito.when(builder.metaCategory(Mockito.any())).thenReturn(builder);
+        Mockito.when(builder.message(Mockito.any())).thenReturn(builder);
+        Mockito.when(builder.build()).thenReturn(notification);
+        Mockito.when(notificationFactory.createBuilder()).thenReturn(builder);
+
+        platformClient.start();
+
+        ArgumentCaptor<Watcher> watcherCaptor = ArgumentCaptor.forClass(Watcher.class);
+        Mockito.verify(op).watch(watcherCaptor.capture());
+        Watcher watcher = watcherCaptor.getValue();
+        MatcherAssert.assertThat(watcher, Matchers.notNullValue());
+
+        watcher.eventReceived(Action.MODIFIED, endpoints);
+
+        ServiceRef serviceRef =
+                new ServiceRef(
+                        connectionToolkit,
+                        address.getIp(),
+                        port.getPort(),
+                        address.getTargetRef().getName());
+
+        Mockito.verify(notificationFactory, Mockito.times(2)).createBuilder();
+        Mockito.verify(builder, Mockito.times(2)).metaCategory("TargetJvmDiscovery");
+
+        InOrder messageOrder = Mockito.inOrder(builder);
+        messageOrder
+                .verify(builder)
+                .message(Map.of("event", Map.of("kind", EventKind.LOST, "serviceRef", serviceRef)));
+        messageOrder
+                .verify(builder)
+                .message(
+                        Map.of("event", Map.of("kind", EventKind.FOUND, "serviceRef", serviceRef)));
+        Mockito.verify(builder, Mockito.times(2)).build();
+        Mockito.verify(notification, Mockito.times(2)).send();
     }
 }
